+++ Assets/js/add-shortcuts.js
@@ -96,6 +96,9 @@ function tagiShortcuts(original) {
 			window.location = '/addshortcuts/v?v=0';
 		});
 		KB.onKey('v+c', function () {
+			window.location = '/addshortcuts/completedLastWeek';
+		});
+		KB.onKey('v+v+c', function () {
 			window.location = '/addshortcuts/completedThisWeek';
 		});