+++ Controller/AddShortcutsController.php
@@ -125,6 +125,38 @@ class AddShortcutsController extends \Kanboard\Controller\PluginController
     }
 
     /**
+     * Get the date range for the last week
+     * and use it in the search.
+     */
+    public function viewCompletedLastWeek()
+    {
+        $url = '/?controller=SearchController&action=index&search=completedRange%3A"{RANGE}"';
+        $today = new \DateTime('today');
+        $interval = \DateInterval::createFromDateString('7 days');
+
+        // get start of week
+        if ($today->format('N') == 1) {
+            $start = $today;
+        } else {
+            $start = new \DateTime('last monday');
+        }
+        $start = $start->sub($interval);
+
+        // get end of week
+        if ($today->format('N') == 7) {
+            $end = $today;
+        } else {
+            $end = new \DateTime('next sunday');
+        }
+        $end = $end->sub($interval);
+
+        $range = $start->format('Y-m-d') . '..' . $end->format('Y-m-d');
+        $url = str_replace('{RANGE}', $range, $url);
+
+        return $this->response->redirect($url, true);
+    }
+
+    /**
      * Show the modal to configure a new custom
      * shortcut with the actual url in the adressbar.
      *