+++ Locale/de_DE/translations.php
@@ -14,5 +14,6 @@ return array(
     'Open task switcher' => 'Aufgaben-Auswahl öffnen',
     'Caption' => 'Bezeichnung',
     'Show tasks, which were completed in the actual week (Mo-Su)' => 'Zeige Aufgaben, die in dieser Woche erlidgt wurden (Mo-So)',
+    'Show tasks, which were completed in the last week (Mo-Su)' => 'Zeige Aufgaben, die in der letzten Woche erlidgt wurden (Mo-So)',
     'Entered data will overwrite chosen shortcut number!' => 'Eingegebene Daten werden gewählte Shortcut-Nummer überschreiben!',
 );